requirements.txt
fastapi
uvicorn
langchain-google-genai
langchain_core
python-dotenv
pydantic
google-cloud-run